OJ-1995 code refactoring
