+++ integration-tests/jest.config-aws.ts
@@ -4,9 +4,8 @@
  */
 import type { Config } from "@jest/types";
 import commonConfig from "./jest.config.common";
-
 const config: Config.InitialOptions = {
   ...commonConfig,
-  testMatch: ["**/tests/**/*.test.ts"],
+  testMatch: ["**/tests/aws/*/*.test.ts"],
 };
 export default config;
