+++ integration-tests/jest.config-mocked.ts
@@ -4,9 +4,8 @@
  */
 import type { Config } from "@jest/types";
 import commonConfig from "./jest.config.common";
-
 const config: Config.InitialOptions = {
   ...commonConfig,
-  testMatch: ["**/tests/**/*.test.ts"],
+  testMatch: ["**/tests/mocked/**/*.test.ts"],
 };
 export default config;
