+++ integration-tests/package.json
@@ -1,14 +1,17 @@
 {
-  "name": "otg-generator-test",
+  "name": "otg-hmrc-api",
   "version": "1.0.0",
   "description": "",
-  "main": "app.js",
   "scripts": {
     "lint:prettier-fix": "pre-commit run prettier --all-files",
-    "lint:eslint": "eslint .",
+    "lint:eslint": "eslint ..",
     "compile": "tsc",
-    "unit": "jest --config=jest.config.ts",
-    "test": "npm run compile && npm run unit"
+    "unit:all": "jest --config=jest.config.ts",
+    "unit:aws": "jest --config=jest.config-aws.ts",
+    "unit:mocked": "jest --config=jest.config-mocked.ts",
+    "test:all": "npm run compile && npm run unit:all",
+    "test:aws": "npm run compile && npm run unit:aws",
+    "test:mocked": "npm run compile && npm run unit:mocked"
   },
   "author": "alphagov",
   "license": "MIT",
@@ -18,11 +21,15 @@
     "esbuild": "0.17.18"
   },
   "devDependencies": {
-    "@aws-sdk/client-sfn": "^3.385.0",
+    "@aws-sdk/client-cloudformation": "^3.405.0",
+    "@aws-sdk/client-dynamodb": "^3.398.0",
+    "@aws-sdk/client-sfn": "^3.405.0",
+    "@aws-sdk/lib-dynamodb": "^3.398.0",
     "@types/aws-lambda": "8.10.114",
     "@types/jest": "^29.5.3",
     "@typescript-eslint/eslint-plugin": "5.10.2",
     "@typescript-eslint/parser": "5.10.2",
+    "aws-testing-library": "^4.0.6",
     "esbuild-jest": "0.5.0",
     "eslint": "8.8.0",
     "eslint-config-prettier": "8.3.0",
