+++ integration-tests/tests/mocked/generate-totp/makefile
@@ -1,6 +1,5 @@
-ROOT_DIR:=$(shell dirname $(realpath $(firstword $(MAKEFILE_LIST))))
 STATE_MACHINE_NAME=oauth-token-generator
-STATE_MACHINE_DEFINITION_FILE=file://${ROOT_DIR}/../step-functions/${STATE_MACHINE_NAME}.asl.json
+STATE_MACHINE_DEFINITION_FILE=file://../../../../step-functions/${STATE_MACHINE_NAME}.asl.json
 STATE_MACHINE_ARN=arn:aws:states:local:123456789012:stateMachine:${STATE_MACHINE_NAME}
 STATE_MACHINE_EXECUTION_ARN=arn:aws:states:local:123456789012:execution:${STATE_MACHINE_NAME}
